% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_integrated}
\alias{pbmc_integrated}
\title{pbmc_integrated}
\format{
A SingleCellExperiment object.
}
\usage{
data(pbmc_integrated)
}
\description{
A SingleCellExperiment dataset of single cell RNA sequencing data. Used for avoiding to wait for calculations
}
\keyword{datasets}
